# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Dipes & Wipes API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /babies:
    get:
      operationId: listBabies
      tags: [babies]
      summary: List all baby profiles
      responses:
        "200":
          description: List of babies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Baby"
    post:
      operationId: createBaby
      tags: [babies]
      summary: Create a baby profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BabyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Baby"
  /babies/{id}:
    get:
      operationId: getBaby
      tags: [babies]
      summary: Get a baby profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Baby
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Baby"
        "404":
          description: Not found
    patch:
      operationId: updateBaby
      tags: [babies]
      summary: Update a baby profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BabyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Baby"
    delete:
      operationId: deleteBaby
      tags: [babies]
      summary: Delete a baby profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Baby:
      type: object
      required: [id, name, birthDate, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        birthDate:
          type: string
          format: date
        gender:
          type: ["string", "null"]
        avatarEmoji:
          type: ["string", "null"]
        currentWeightLbs:
          type: ["number", "null"]
        currentHeightIn:
          type: ["number", "null"]
        weightPercentile:
          type: ["number", "null"]
        heightPercentile:
          type: ["number", "null"]
        currentHeightPercentile:
          type: ["number", "null"]
        currentDiaperSize:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    BabyInput:
      type: object
      required: [name, birthDate]
      properties:
        name:
          type: string
          minLength: 1
        birthDate:
          type: string
          format: date
        gender:
          type: string
        avatarEmoji:
          type: string
        currentWeightLbs:
          type: number
        currentHeightIn:
          type: number
        weightPercentile:
          type: number
        heightPercentile:
          type: number
        notes:
          type: string
    BabyUpdate:
      type: object
      properties:
        name:
          type: string
        birthDate:
          type: string
          format: date
        gender:
          type: string
        avatarEmoji:
          type: string
        currentWeightLbs:
          type: number
        currentHeightIn:
          type: number
        weightPercentile:
          type: number
        heightPercentile:
          type: number
        currentDiaperSize:
          type: string
        notes:
          type: string
...